route protection, header for subscription status; refresh showing ads in dashboard and summary
