+++ src/app/add-ads-account/AddAdsAccountWrapper.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { ProtectedRoute } from "@/components/auth";
+import { AddAdsAccount } from "./AddAdsAccount";
+
+export default function AddAdsAccountWrapper() {
+  return (
+    <ProtectedRoute>
+      <AddAdsAccount />
+    </ProtectedRoute>
+  );
+}
