+++ src/app/dashboard/DashboardWrapper.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { ProtectedRoute } from "@/components/auth";
+import Dashboard from "./Dashboard";
+
+export default function DashboardWrapper() {
+  return (
+    <ProtectedRoute>
+      <Dashboard />
+    </ProtectedRoute>
+  );
+}
