+++ src/app/settings/settings/ad-accounts/AdAccountsWrapper.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { ProtectedRoute } from "@/components/auth";
+import AdAccounts from "./AdAccounts";
+
+export default function AdAccountsWrapper() {
+  return (
+    <ProtectedRoute>
+      <AdAccounts />
+    </ProtectedRoute>
+  );
+}
