+++ src/app/settings/settings/alerts/AlertsWrapper.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { ProtectedRoute } from "@/components/auth";
+import Alerts from "./Alerts";
+
+export default function AlertsWrapper() {
+  return (
+    <ProtectedRoute>
+      <Alerts />
+    </ProtectedRoute>
+  );
+}
