+++ src/app/settings/settings/users/UsersWrapper.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { ProtectedRoute } from "@/components/auth";
+import Users from "./Users";
+
+export default function UsersWrapper() {
+  return (
+    <ProtectedRoute>
+      <Users />
+    </ProtectedRoute>
+  );
+}
