+++ src/app/summary/SummaryWrapper.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { ProtectedRoute } from "@/components/auth";
+import Summary from "./Summary";
+
+export default function SummaryWrapper() {
+  return (
+    <ProtectedRoute>
+      <Summary />
+    </ProtectedRoute>
+  );
+}
