+++ src/components/auth/index.ts
@@ -1,0 +1,3 @@
+export { ProtectedRoute } from "./ProtectedRoute";
+export { AuthStateHandler } from "./auth-state-handler";
+export { AuthSync } from "./auth-sync";
